move class AppendInfo to logger.py
